Beginning of a large amount of work.
Updated database schema with the new models we're going to go with for
now. Started the lengthy process of isolating all database interactions
to util.py (later to be named client.py with a database client class).
Updated some views. Still need to fix crazy stuff in imports but I'm
tired.